Modify the way of portal login
