+++ crawler/portal/portal.go
@@ -2,6 +2,7 @@ package portal
 
 import (
 	"bytes"
+	"encoding/json"
 	"log"
 	"net/http"
 	"net/http/cookiejar"
@@ -10,16 +11,29 @@ import (
 )
 
 /*
+Result is the result of Login response
+success: is login successed
+status: the status of response
+data: the login result
+*/
+type Result struct {
+	Success bool
+	Status  int
+	Data    interface{}
+}
+
+/*
 Login is a function which handle login request
 return: http.Client for future reuse
 studentID: the id of student
 password: the password of student
 */
-func Login(client *http.Client, studentID string, password string) (*http.Response, error) {
+func Login(client *http.Client, studentID string, password string) (loginResult Result, err error) {
 	req := newRequest(studentID, password)
 	resp, err := client.Do(req)
+	loginResult = handleResponse(resp)
 
-	return resp, err
+	return loginResult, err
 }
 
 /*
@@ -33,6 +47,24 @@ func NewClient() *http.Client {
 	}
 
 	return client
+}
+
+func handleResponse(resp *http.Response) (loginResult Result) {
+	defer resp.Body.Close()
+
+	var data map[string]interface{}
+
+	json.NewDecoder(resp.Body).Decode(&data)
+
+	statusCode := 200
+	isSuccess := data["success"].(bool)
+	message := "登入成功"
+	if !isSuccess {
+		statusCode = 401
+		message = "帳號或密碼錯誤，請重新輸入。"
+	}
+
+	return Result{Success: isSuccess, Status: statusCode, Data: message}
 }
 
 func newRequest(studentID string, password string) *http.Request {
